Add analytics to articles
